+++ sim/verilator/testb.h
@@ -49,7 +49,7 @@ template <class VA>	class TESTB {
 public:
 	VA		*m_core;
 	VerilatedVcdC*	m_trace;
-	unsigned long	m_tickcount;
+	uint64_t	m_tickcount;
 
 	TESTB(void) : m_trace(NULL), m_tickcount(0l) {
 		m_core = new VA;
@@ -92,14 +92,14 @@ public:
 		// logic depends.  This forces that logic to be recalculated
 		// before the top of the clock.
 		eval();
-		if (m_trace) m_trace->dump(10*m_tickcount-2);
+		if (m_trace) m_trace->dump((uint64_t)(10*m_tickcount-2));
 		m_core->i_clk = 1;
 		eval();
-		if (m_trace) m_trace->dump(10*m_tickcount);
+		if (m_trace) m_trace->dump((uint64_t)(10*m_tickcount));
 		m_core->i_clk = 0;
 		eval();
 		if (m_trace) {
-			m_trace->dump(10*m_tickcount+5);
+			m_trace->dump((uint64_t)(10*m_tickcount+5));
 			m_trace->flush();
 		}
 	}
